requirements.txt
Adafruit-Blinka==6.4.1
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-motor==3.3.0
adafruit-circuitpython-pca9685==3.3.5
adafruit-circuitpython-register==1.9.5
adafruit-circuitpython-servokit==1.3.3
Adafruit-PlatformDetect==3.4.1
Adafruit-PureIO==1.1.8
paho-mqtt==1.5.1
pkg-resources==0.0.0
pyftdi==0.52.9
pyserial==3.5
pyusb==1.1.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
